src/requirements.txt
Flask==3.1.2
pandas==2.3.3
openpyxl==3.1.5
fpdf2==2.8.5
PyJWT==2.10.1
python-dotenv==1.2.1
flasgger==0.9.7.1
